Avoid caching missing partial CI rollups
